image_processor/ai_utils.py
import time
import os
from PIL import Image, ImageFilter, ImageEnhance
from django.core.files.base import ContentFile
from django.core.files.storage import default_storage
import io


def mock_cartoon_processing(image_path, output_path):
    """
    Mock AI processing function that applies cartoon-like effects.
    In production, replace this with actual AI model calls (Toonify, AnimeGAN, etc.)
    """
    try:
        # Simulate processing time
        processing_start = time.time()
        
        # Open the original image
        with Image.open(image_path) as img:
            # Convert to RGB if necessary
            if img.mode != 'RGB':
                img = img.convert('RGB')
            
            # Apply cartoon-like effects (mock processing)
            # 1. Enhance colors
            enhancer = ImageEnhance.Color(img)
            img = enhancer.enhance(1.5)  # Increase saturation
            
            # 2. Apply slight blur and then sharpen for cartoon effect
            img = img.filter(ImageFilter.SMOOTH_MORE)
            img = img.filter(ImageFilter.SHARPEN)
            
            # 3. Enhance contrast
            enhancer = ImageEnhance.Contrast(img)
            img = enhancer.enhance(1.3)
            
            # 4. Apply a slight gaussian blur for smoothing
            img = img.filter(ImageFilter.GaussianBlur(radius=0.5))
            
            # Save the processed image
            img.save(output_path, 'JPEG', quality=95)
        
        processing_time = time.time() - processing_start
        return True, processing_time, None
        
    except Exception as e:
        return False, 0, str(e)


def process_image_to_cartoon(image_processing_obj):
    """
    Process an ImageProcessing object to create cartoon version
    """
    try:
        # Update status to processing
        image_processing_obj.status = 'processing'
        image_processing_obj.save()
        
        # Get the original image path
        original_path = image_processing_obj.original_image.path
        
        # Create output filename
        original_name = os.path.basename(original_path)
        name, ext = os.path.splitext(original_name)
        output_filename = f"{name}_cartoon{ext}"
        
        # Create temporary output path
        temp_output_path = os.path.join(
            os.path.dirname(original_path), 
            f"temp_{output_filename}"
        )
        
        # Process the image
        success, processing_time, error_msg = mock_cartoon_processing(
            original_path, 
            temp_output_path
        )
        
        if success:
            # Save the processed image to the model
            with open(temp_output_path, 'rb') as f:
                image_processing_obj.processed_image.save(
                    output_filename,
                    ContentFile(f.read()),
                    save=False
                )
            
            # Update status and processing time
            image_processing_obj.status = 'completed'
            image_processing_obj.processing_time = processing_time
            image_processing_obj.save()
            
            # Clean up temporary file
            if os.path.exists(temp_output_path):
                os.remove(temp_output_path)
            
            return True, None
        else:
            # Update status to failed
            image_processing_obj.status = 'failed'
            image_processing_obj.error_message = error_msg
            image_processing_obj.save()
            
            return False, error_msg
            
    except Exception as e:
        # Update status to failed
        image_processing_obj.status = 'failed'
        image_processing_obj.error_message = str(e)
        image_processing_obj.save()
        
        return False, str(e)


# Future integration functions for real AI models
def integrate_toonify_api(image_path, api_key):
    """
    Placeholder for Toonify API integration
    """
    # TODO: Implement actual Toonify API calls
    pass


def integrate_animegan_model(image_path, model_path):
    """
    Placeholder for AnimeGAN model integration
    """
    # TODO: Implement actual AnimeGAN model processing
    pass
