requirements.txt
Django>=4.2.7
Pillow>=9.0.0
python-decouple==3.8
whitenoise>=6.6.0
gunicorn>=21.2.0

# Optional: For background processing
# celery>=5.3.0
# redis>=5.0.0
